Extracts queue classes from channel module
