+++ src/simperium/channel.js
@@ -3,6 +3,8 @@ import { format, inherits } from 'util'
 import { EventEmitter } from 'events'
 import { parseMessage, parseVersionMessage, change as change_util } from './util'
 import { v4 as uuid } from 'uuid'
+import NetworkQueue from './queues/network-queue';
+import LocalQueue from './queues/local-queue';
 
 const UNKNOWN_CV = '?';
 const CODE_INVALID_VERSION = 405;
@@ -342,7 +344,6 @@ internal.indexingComplete = function() {
  * @returns {Promise<Void>} - resolves once the change version is saved
  */
 
-
 /**
  * Maintains syncing state for a Simperium bucket.
  *
@@ -661,199 +662,6 @@ Channel.prototype.onVersion = function( data ) {
 	this.emit( 'version.' + ghost.id, ghost.id, ghost.version, ghost.data );
 	this.emit( 'version.' + ghost.id + '.' + ghost.version, ghost.data );
 };
-
-function NetworkQueue() {
-	this.queues = {};
-}
-
-NetworkQueue.prototype.queueFor = function( id ) {
-	var queues = this.queues,
-		queue = queues[id];
-
-	if ( !queue ) {
-		queue = new Queue();
-		queue.on( 'finish', function() {
-			delete queues[id];
-		} );
-		queues[id] = queue;
-	}
-
-	return queue;
-};
-
-function Queue() {
-	this.queue = [];
-	this.running = false;
-}
-
-inherits( Queue, EventEmitter );
-
-// Add a function at the end of the queue
-Queue.prototype.add = function( fn ) {
-	this.queue.push( fn );
-	this.start();
-	return this;
-};
-
-Queue.prototype.start = function() {
-	if ( this.running ) return;
-	this.running = true;
-	this.emit( 'start' );
-	setImmediate( this.run.bind( this ) );
-}
-
-Queue.prototype.run = function() {
-	var fn;
-	this.running = true;
-
-	if ( this.queue.length === 0 ) {
-		this.running = false;
-		this.emit( 'finish' );
-		return;
-	}
-
-	fn = this.queue.shift();
-	fn( this.run.bind( this ) );
-}
-
-function LocalQueue( store ) {
-	this.store = store;
-	this.sent = {};
-	this.queues = {};
-	this.ready = false;
-}
-
-inherits( LocalQueue, EventEmitter );
-
-LocalQueue.prototype.start = function() {
-	var queueId;
-	this.ready = true;
-	for ( queueId in this.queues ) {
-		this.processQueue( queueId );
-	}
-}
-
-LocalQueue.prototype.pause = function() {
-	this.ready = false;
-};
-
-LocalQueue.prototype.acknowledge = function( change ) {
-	if ( this.sent[change.id] === change ) {
-		delete this.sent[change.id];
-	}
-
-	this.processQueue( change.id );
-}
-
-LocalQueue.prototype.queue = function( change ) {
-	var queue = this.queues[change.id];
-
-	if ( !queue ) {
-		queue = [];
-		this.queues[change.id] = queue;
-	}
-
-	queue.push( change );
-
-	this.emit( 'queued', change.id, change, queue );
-
-	if ( !this.ready ) return;
-
-	this.processQueue( change.id );
-};
-
-LocalQueue.prototype.hasChanges = function() {
-	return Object.keys( this.queues ).length > 0;
-};
-
-LocalQueue.prototype.dequeueChangesFor = function( id ) {
-	var changes = [], sent = this.sent[id], queue = this.queues[id];
-
-	if ( sent ) {
-		delete this.sent[id];
-		changes.push( sent );
-	}
-
-	if ( queue ) {
-		delete this.queues[id];
-		changes = changes.concat( queue );
-	}
-
-	return changes;
-};
-
-LocalQueue.prototype.processQueue = function( id ) {
-	var queue = this.queues[id];
-	var compressAndSend = this.compressAndSend.bind( this, id );
-
-	// there is no queue, don't do anything
-	if ( !queue ) return;
-
-	// queue is empty, delete it from memory
-	if ( queue.length === 0 ) {
-		delete this.queues[id];
-		return;
-	}
-
-	// waiting for a previous sent change to get acknowledged
-	if ( this.sent[id] ) {
-		this.emit( 'wait', id );
-		return;
-	}
-
-	this.store.get( id ).then( compressAndSend );
-}
-
-LocalQueue.prototype.compressAndSend = function( id, ghost ) {
-	var changes = this.queues[id];
-	var change;
-	var target = ghost.data;
-	var c;
-	var type;
-
-	// a change was sent before we could compress and send
-	if ( this.sent[id] ) {
-		this.emit( 'wait', id );
-		return;
-	}
-
-	if ( changes.length === 1 ) {
-		change = changes.shift();
-		this.sent[id] = change;
-		this.emit( 'send', change );
-		return;
-	}
-
-	if ( changes.length > 1 && changes[0].type === change_util.type.REMOVE ) {
-		change = changes.shift();
-		changes.splice( 0, changes.length - 1 );
-		this.sent[id] = change;
-		this.emit( 'send', change );
-	}
-
-	while ( changes.length > 0 ) {
-		c = changes.shift();
-
-		if ( c.o === change_util.type.REMOVE ) {
-			changes.unshift( c );
-			break;
-		}
-
-		target = change_util.apply( c.v, target );
-	}
-
-	type = target === null ? change_util.type.REMOVE : change_util.type.MODIFY;
-	change = change_util.buildChange( type, id, target, ghost );
-
-	this.sent[id] = change;
-	this.emit( 'send', change );
-}
-
-LocalQueue.prototype.resendSentChanges = function() {
-	for ( let ccid in this.sent ) {
-		this.emit( 'send', this.sent[ccid] )
-	}
-}
 
 /**
  * Since revision data is basically immutable we can prevent the
